test: category all GAgentBase tests.
