+++ src/AISmart.Application.Grains/Agents/MarketLeader/MarketLeaderGAgent.cs
@@ -11,7 +11,7 @@ namespace AISmart.Application.Grains.Agents.MarketLeader;
 [Description("Marketing department，I can handle tasks related to the marketing department.")]
 [StorageProvider(ProviderName = "PubSubStore")]
 [LogConsistencyProvider(ProviderName = "LogStorage")]
-public class MarketLeaderGAgent : GAgentBase<MarketLeaderAgentState, MarketLeaderGEvent>
+public class MarketLeaderGAgent : EventHandler<MarketLeaderAgentState, MarketLeaderGEvent>
 {
     public MarketLeaderGAgent(ILogger<MarketLeaderGAgent> logger) : base(logger)
     {
