+++ src/AISmart.Application.Grains/Agents/Publisher/PublishingGAgent.cs
@@ -15,7 +15,7 @@ public class PublishingAgentState : StateBase
 
 [StorageProvider(ProviderName = "PubSubStore")]
 [LogConsistencyProvider(ProviderName = "LogStorage")]
-public class PublishingGAgent : GAgentBase<PublishingAgentState, PublishingGEvent>, IPublishingGAgent
+public class PublishingGAgent : EventHandler<PublishingAgentState, PublishingGEvent>, IPublishingGAgent
 {
     public PublishingGAgent(ILogger<PublishingGAgent> logger) : base(logger)
     {
